Promocodes,home and clean up

+++ src/app/admin/products/products.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { Product } from 'src/app/shared/classes/product';
+import { AuthService } from 'src/app/shared/services/auth.service';
 import { ModalService } from 'src/app/shared/services/modal.service';
 import { ProductService } from 'src/app/shared/services/product.service';
 
@@ -47,7 +48,8 @@ export class AdminProductsComponent implements OnInit {
 
 	constructor(private productsService:ProductService,
 		private modalService:ModalService,
-		private toast:ToastrService) { }
+		private toast:ToastrService,
+		private authService:AuthService) { }
 
 	ngOnInit(): void {
 		this.newProductForm = new FormGroup({
@@ -89,25 +91,42 @@ export class AdminProductsComponent implements OnInit {
 
 
 	addNewProduct(){
-		if(this.newProductForm.valid && this.productImages.length >0){
-			this.productsService.addProduct(
-				this.newProductForm.get('productName').value,
-				this.newProductForm.get('productColor').value.value,
-				this.productImages[0] ,
-				this.newProductForm.get('productDescription').value,
-				this.newProductForm.get('productPrice').value,
-				this.newProductForm.get('oldPrice').value,
-				this.newProductForm.get('category').value.value,
-			).subscribe((res:any) =>{
-				if(res && res.product){
-					if(this.productImages.length>1){
-						this.productsService.addProductImages(
-							res.product.productId,
-							this.productImages.slice(1)
-						).subscribe((imgRes:any) =>{
-							if(imgRes && imgRes.images){
+		this.authService.newUser.subscribe(user =>{
+			if(user && user.userId && user.type == 'admin'){
+				if(this.newProductForm.valid && this.productImages.length >0){
+					this.productsService.addProduct(
+						this.newProductForm.get('productName').value,
+						this.newProductForm.get('productColor').value.value,
+						this.productImages[0] ,
+						this.newProductForm.get('productDescription').value,
+						this.newProductForm.get('productPrice').value,
+						this.newProductForm.get('oldPrice').value,
+						this.newProductForm.get('category').value.value,
+					).subscribe((res:any) =>{
+						if(res && res.product){
+							if(this.productImages.length>1){
+								this.productsService.addProductImages(
+									res.product.productId,
+									this.productImages.slice(1)
+								).subscribe((imgRes:any) =>{
+									if(imgRes && imgRes.images){
+										this.toast.success('Added Successfully')
+										this.getAllProducts();
+										this.isModalVisible = false;
+										this.newProductForm = new FormGroup({
+											productName: new FormControl(null,Validators.required),
+											productPrice: new FormControl(null,[Validators.required,Validators.pattern(/^[0-9]+$/)]),
+											oldPrice: new FormControl(null,[Validators.required,Validators.pattern(/^[0-9]+$/)]),
+											productColor: new FormControl(null,Validators.required),
+											
+											productDescription: new FormControl(null,Validators.required),
+											category: new FormControl(null,Validators.required),
+										})
+										this.productImages = []
+									}
+								})
+							}else{
 								this.toast.success('Added Successfully')
-								this.getAllProducts();
 								this.isModalVisible = false;
 								this.newProductForm = new FormGroup({
 									productName: new FormControl(null,Validators.required),
@@ -120,26 +139,14 @@ export class AdminProductsComponent implements OnInit {
 								})
 								this.productImages = []
 							}
-						})
-					}else{
-						this.toast.success('Added Successfully')
-						this.isModalVisible = false;
-						this.newProductForm = new FormGroup({
-							productName: new FormControl(null,Validators.required),
-							productPrice: new FormControl(null,[Validators.required,Validators.pattern(/^[0-9]+$/)]),
-							oldPrice: new FormControl(null,[Validators.required,Validators.pattern(/^[0-9]+$/)]),
-							productColor: new FormControl(null,Validators.required),
-							
-							productDescription: new FormControl(null,Validators.required),
-							category: new FormControl(null,Validators.required),
-						})
-						this.productImages = []
-					}
+						}
+					})
+				}else{
+					this.toast.error('Please complete the missing data')
 				}
-			})
-		}else{
-			this.toast.error('Please complete the missing data')
-		}
+			}
+		})
+		
 	}
 
 	onEditProduct(product){
@@ -187,68 +194,83 @@ export class AdminProductsComponent implements OnInit {
 		
 	}
 	getProductById(productId){
-		this.productsService.getProductById(productId).subscribe((res)=>{
-			if(res && res.product){
-				this.selectedProductImages = res.product.productImages;
+		this.authService.newUser.subscribe(user =>{
+			if(user && user.userId && user.type == 'admin'){
+				this.productsService.getProductById(productId).subscribe((res)=>{
+					if(res && res.product){
+						this.selectedProductImages = res.product.productImages;
+					}
+				})
 			}
 		})
+		
 	}
 
 	 updateProduct(){
-		if(this.updateProductForm.valid){
-				this.productsService.editProduct(
-					this.selectedProduct.productId,
-					this.updateProductForm.get('productName').value,
-					this.updateProductForm.get('productColor').value.value,
-					this.updateProductForm.get('productDescription').value,
-					this.updateProductForm.get('productPrice').value,
-					this.updateProductForm.get('oldPrice').value,
-					this.editProductImages[0],
-				).subscribe((res:any) =>{
-					if(res && res.message == "product Updated succssefully"){
-						if(this.editProductImages.length >1){
-							this.productsService.addProductImages(
-								this.selectedProduct.productId,
-								this.editProductImages.slice(1)
-							).subscribe((imgRes:any) =>{
-								if(imgRes && imgRes.images){
-									this.toast.success('Updated Successfully')
-									this.getAllProducts();
-									this.editProductImages = []
-									this.isEditModalVisible = false;
-								}
-							})
+		this.authService.newUser.subscribe(user =>{
+			if(user && user.userId && user.type == 'admin'){
+				if(this.updateProductForm.valid){
+					this.productsService.editProduct(
+						this.selectedProduct.productId,
+						this.updateProductForm.get('productName').value,
+						this.updateProductForm.get('productColor').value.value,
+						this.updateProductForm.get('productDescription').value,
+						this.updateProductForm.get('productPrice').value,
+						this.updateProductForm.get('oldPrice').value,
+						this.editProductImages[0],
+					).subscribe((res:any) =>{
+						if(res && res.product){
+							if(this.editProductImages.length >1){
+								this.productsService.addProductImages(
+									this.selectedProduct.productId,
+									this.editProductImages.slice(1)
+								).subscribe((imgRes:any) =>{
+									if(imgRes && imgRes.images){
+										this.toast.success('Updated Successfully')
+										this.getAllProducts();
+										this.editProductImages = []
+										this.isEditModalVisible = false;
+									}
+								})
+							}else{
+								this.toast.success('Updated Successfully')
+								this.getAllProducts();
+								this.editProductImages = []
+								this.isEditModalVisible = false;
+							}
+							
 						}else{
 							this.toast.success('Updated Successfully')
 							this.getAllProducts();
-							this.editProductImages = []
 							this.isEditModalVisible = false;
+							this.editProductImages = []
 						}
-						
-					}else{
-						this.toast.success('Updated Successfully')
-						this.getAllProducts();
-						this.isEditModalVisible = false;
-						this.editProductImages = []
-					}
-				})
-			
-		}else{
-			this.toast.error('Please complete the missing data')
-		}
+					})
+				
+			}else{
+				this.toast.error('Please complete the missing data')
+			}
+			}
+		})
+		
 	 }
 
 
 
 	getAllProducts(){
-		this.productsService.getAllProducts().subscribe(res =>{
-			this.products = []
-			if(res.products){
-				this.products = res.products
-			}else{
-				this.products = []
+		this.authService.newUser.subscribe(user =>{
+			if(user && user.userId && user.type == 'admin'){
+				this.productsService.getAllProducts().subscribe(res =>{
+					this.products = []
+					if(res.products){
+						this.products = res.products
+					}else{
+						this.products = []
+					}
+				})
 			}
 		})
+		
 	}
 
 	openProductModal(product?){
@@ -256,22 +278,30 @@ export class AdminProductsComponent implements OnInit {
 	}
 
 	deleteProduct(product){
-		this.productsService.removeProduct(product.productId).subscribe((res:any) =>{
-			if(res && res.message =="product deleted succssefully"){
-				this.toast.success('product deleted succssefully');
-				this.getAllProducts();
+		this.authService.newUser.subscribe(user =>{
+			if(user && user.userId && user.type == 'admin'){
+				this.productsService.removeProduct(product.productId).subscribe((res:any) =>{
+					if(res && res.product){
+						this.toast.success('product deleted succssefully');
+						this.getAllProducts();
+					}
+				})
 			}
 		})
 	}
 
 	deleteImage(image){
-		console.log(image)
-		this.productsService.removeOneProductImage(image.imageId).subscribe((res:any) =>{
-			if(res && res.message == 'image Delete successfully'){
-				this.toast.success('Image deleted succssefully');
-				this.getProductById(image.product_id);
+		this.authService.newUser.subscribe(user =>{
+			if(user && user.userId && user.type == 'admin'){
+				this.productsService.removeOneProductImage(image.imageId).subscribe((res:any) =>{
+					if(res && res.message == 'image Delete successfully'){
+						this.toast.success('Image deleted succssefully');
+						this.getProductById(image.product_id);
+					}
+				})
 			}
 		})
+		
 	}
 
 }
